+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Allow to pass `--version` in format `major.minor.patch[-count][-hash]`
 
+### Changed
+
+- RPM header: added `PAYLOADDIGEST` and `PAYLOADDIGESTALGO` flags,
+  removed `RPMVERSION`.
+
 ## [1.3.0] - 2020-01-13
 
 ### Added